Add rock-proxy route and URL validation
Introduce a dedicated server-side Rock proxy and URL validation.

- Add ~/lib/.server/validate-rock-url.ts to validate and restrict incoming `url` query params (HTTPS + allowed host).
- Create a rock-proxy endpoint (fetchRockProxyHtml + loader) that fetches Rock RMS HTML, rewrites asset URLs, injects styles/scripts, and returns embeddable HTML for iframes.
- Replace the old monolithic rock-page server loader with a lighter rock-page route/loader/meta that simply renders the RockProxyEmbed component (which now points to /rock-proxy when using the advanced proxy).
- Update RockProxyEmbed to use /rock-proxy instead of /rock-page.
- Update netlify.toml to apply X-Frame-Options override to /rock-proxy.

These changes separate concerns (validation, fetching/rewrite, and UI embedding), improve security by centralizing URL validation, and fix the proxy route used by the embed component.
